requirements.txt
bcrypt==4.0.1
cloudinary>=1.44.1
python-dotenv>=1.2.1
fastapi>=0.128.0
google-api-python-client>=2.189.0
google-auth-httplib2>=0.3.0
google-auth-oauthlib>=1.2.4
python-jose[cryptography]>=3.3.0
numpy>=2.4.1
openpyxl>=3.1.5
pandas>=3.0.0
passlib[bcrypt]>=1.7.4
pillow>=12.1.0
psycopg2-binary>=2.9.11
pypdf2>=3.0.1
python-docx>=1.2.0
python-multipart>=0.0.22
reportlab>=4.4.9
sqlalchemy>=2.0.46
supabase>=2.27.3
uvicorn>=0.40.0
